Fix #12:문제 가이드, 피드백 페이지 개선

+++ backend/app/schemas/guide.py
@@ -10,7 +10,7 @@ class ProblemDetailResponse(BaseModel):
     time_limit: int  # milliseconds
     memory_limit: int  # KB
     submission_count: int
-    accepted_count: int
+    accepted_count: float
     solved_count: int
     success_rate: float
     tier: int
